Use SRI vectors only on net7 and later
